ros_test_task-master/src/requirements.txt
fastapi==0.101.1
pydantic_core==2.4.0
uvicorn
requests
